+++ R/fn_exp_categorical_viz.R
@@ -2,10 +2,10 @@
 #'
 #'
 #' @description This function automatically scans through each variable and creates bar plot for categorical variable.
-#' @usage ExpCatViz(data,gp=NULL,fname=NULL,clim=10,col=NULL,
-#' margin=1,Page=NULL,Flip=F,sample=NULL,rdata=FALSE,value=NULL)
+#' @usage ExpCatViz(data,target=NULL,fname=NULL,clim=10,col=NULL,margin=1,Page=NULL,
+#' Flip=F,sample=NULL,rdata=FALSE,value=NULL,gtitle=NULL,theme="Default")
 ##' @param data dataframe or matrix
-##' @param gp target variable. This is not a mandatory field.
+##' @param target target variable. This is not a mandatory field.
 ##' @param fname output file name. Output will be generated in PDF format
 ##' @param clim maximum categories to be considered to include in bar graphs.
 ##' @param col define the colors to fill the bars, default it will take sample colours.
@@ -15,6 +15,8 @@
 ##' @param sample random selection of categorical variable
 ##' @param rdata to plot bar graph for frequency/aggregated table
 ##' @param value value coloumn name. This is mandatory if 'rdata' is TRUE
+##' @param gtitle graph title
+##' @param theme extra themes, geoms, and scales for 'ggplot2' (use theme from ggthemes package)
 ##' @seealso \code{\link[ggplot2:geom_bar]{geom_bar}}
 ##' @return This function returns collated graphs in grid format in PDF or JPEG format. All the files will be stored in the working directory
 ##'
@@ -25,25 +27,32 @@
 ##'Stacked Bar graph by target variable
 ##'
 ##' @examples
-#' ExpCatViz(data=mtcars,gp=NULL,fname=file.path(tempdir(),"Cat_1"),clim=10,margin=1,Page = c(2,2))
-#' ## Generate Bar graph for all the descrete data with column based proportions - random colors
-#' set.seed(1234)
-#' ExpCatViz(data=mtcars,gp="gear",fname=file.path(tempdir(),"Cat_2"),clim=10,margin=2,Page = c(2,2))
 #' ## Bar graph for specified variable
 #' mtdata <- mtcars
 #' mtdata$carname <- rownames(mtcars)
-#' ExpCatViz(data=mtdata,gp="carname",col="blue",rdata=TRUE,value="mpg")
-##' @importFrom grDevices colors
+#' ExpCatViz(data=mtdata,target="carname",col="blue",rdata=TRUE,value="mpg")
+##' @importFrom grDevices colors pdf dev.off
 ##' @importFrom gridExtra marrangeGrob
+##' @importFrom sampling srswor
 ##' @export ExpCatViz
 
-ExpCatViz <- function(data,gp=NULL,fname=NULL,clim=10,col=NULL,margin=1,Page=NULL,Flip=F,sample=NULL,rdata=FALSE,value=NULL) {
-
+ExpCatViz <- function(data,target=NULL,fname=NULL,clim=10,col=NULL,margin=1,Page=NULL,Flip=F,sample=NULL,rdata=FALSE,value=NULL,gtitle=NULL,theme="Default") {
+  if(!is.data.frame(data)) stop("'data must be a data.frame'")
   if(is.null(data))stop("There is no input data frame")
 
   r<-0
   xx = as.data.frame(data)
-  if(isFALSE(rdata)){
+
+  if(theme=="Default") {
+    mytheme <- theme(axis.text.x=element_text(angle=90,vjust=.5,hjust=0.95,size=8,colour='grey20'),
+                     axis.text.y=element_text(vjust=.5,hjust=0.95,size=8,colour='grey20'),
+                     plot.title = element_text(hjust = 0.5, face = "bold", colour = "#5F9EA0", size = 12))+
+      theme(axis.line = element_line(size=1, colour = "black"),
+            panel.grid.major = element_line(colour = "#d3d3d3",linetype = "dashed"), panel.grid.minor = element_blank(),
+            panel.border = element_blank(), panel.background = element_blank())
+  } else {mytheme <- theme }
+
+  if(rdata==FALSE){
     num_var = names(xx)[sapply(xx, is.numeric)]
     Cat_var <- c(names(xx)[sapply(xx, is.character)],names(xx)[sapply(xx, is.factor)])
 
@@ -54,22 +63,21 @@ ExpCatViz <- function(data,gp=NULL,fname=NULL,clim=10,col=NULL,margin=1,Page=NUL
     if((length(num_var)+length(Cat_var))==0) stop("there is no categorical variable in the data")
 
     wrap_40 <- wrap_format(40)
-
-    if(!is.null(gp)) {
-      # Tar_var = xx[,gp]
+  if(!is.null(target)) {
+      # Tar_var = xx[,target]
       # if(is.numeric(Tar_var)&length(unique(Tar_var))>5)
       #
-      Yvar = as.character(paste0(xx[,gp]))
+      Yvar = as.character(paste0(xx[,target]))
       nlevel = length(unique(Yvar))
       if(is.null(col) & nlevel==1){fill_1 <- c("#5F9EA0")} else
         if(is.null(col) & nlevel==2){fill_1 <- c("#5F9EA0", "#E1B378")} else
-          if(is.null(col) & nlevel>2){fill_1 <- sample(colors(),nlevel)} else
+          if(is.null(col) & nlevel>2){fill_1 <- colors()[srswor(nlevel,length(colors()))==1]} else
             if(!is.null(col) & nlevel>2){fill_1 <- col} else {fill_1 <- col}
       if(length(Cat_var) >0){
         Cat_varlst <- Cat_var[sapply(xx[,Cat_var], function(x){length(unique(x))<=clim & length(unique(x))>=2})]
         Cat_varlst <- c(Cat_varlst,num_var)
-        Cat_varlst <- Cat_varlst[!(Cat_varlst %in% gp)]} else
-        {Cat_varlst <- num_var[!(num_var %in% gp)]}
+        Cat_varlst <- Cat_varlst[!(Cat_varlst %in% target)]} else
+        {Cat_varlst <- num_var[!(num_var %in% target)]}
 
     }
     else {
@@ -84,52 +92,52 @@ ExpCatViz <- function(data,gp=NULL,fname=NULL,clim=10,col=NULL,margin=1,Page=NUL
     }
     options(scipen = 100)
 
-    if(!is.null(sample)) {
-      if(sample>length(Cat_varlst)) {sample <- length(Cat_varlst)} else
-      {Cat_varlst <- Cat_varlst[sample(1:length(Cat_varlst),sample)]}}
-  } else {
+    if(!is.null(sample))
+      {
+      if(sample>length(Cat_varlst)) {Cat_varlst <- Cat_varlst} else {
+        Cat_varlst <- Cat_varlst[srswor(sample,length(Cat_varlst))==1]
+      }
+              }
+  } ## Bar plot ends
+
+  else
+  {
     if(is.null(value)) stop("value column is missing")
     Cat_varlst <- value
   }
 
+  ### GGPLOT graph start from here
   plot_l <- lapply(Cat_varlst, function(j){
     Xvar = as.character(paste0(xx[,j]))
-
-    if (isTRUE(rdata)){
+    if(!is.null(gtitle)) {chart_title <- gtitle} else chart_title<-NULL
+    if (rdata==TRUE){
       data = as.data.frame(data)
-      tbl <- data[,c(gp,value)]
+      tbl <- data[,c(target,value)]
       names(tbl) <- c("Xvar","Freq")
       tbl$Freq <- round(tbl$Freq,1)
       nlevel <- length(unique(tbl$Xvar))
       if(is.null(col)& nlevel==1){fill_1 <- c("#5F9EA0")}
       else
-        if(is.null(col) & nlevel>=2){fill_1 <- sample(colors(),nlevel)}
+        if(is.null(col) & nlevel>=2){fill_1 <- colors()[srswor(nlevel,length(colors()))==1]}
       else
         fill_1=col
 
       pp = ggplot(tbl, aes(y=Freq, x=reorder(Xvar,Freq),label = paste0(Freq))) +
-        geom_bar( stat="identity", position="dodge",fill=fill_1)+xlab("")+
-        ylab(value)+
-        geom_text(size = 3, position=position_dodge(width=0.9),vjust=0.8)+
+        geom_bar( stat="identity", position="dodge",fill=fill_1)+xlab("")+ggtitle(chart_title)+
+        ylab(value)+ geom_text(size = 3, position=position_dodge(width=0.9),vjust=0)+
         scale_x_discrete(labels = wrap_format(8))+
-        scale_y_continuous(labels = dollar_format(suffix = "", prefix = ""))+
-        theme(axis.text.x=element_text(angle=90,vjust=.5,hjust=0.95,size=8,colour='grey20'),
-              axis.text.y=element_text(vjust=.5,hjust=0.95,size=8,colour='grey20'),
-              plot.title = element_text(hjust = 0.5, face = "bold", colour = "#5F9EA0", size = 12))+
-        theme(axis.line = element_line(size=1, colour = "black"),
-              panel.grid.major = element_line(colour = "#d3d3d3",linetype = "dashed"), panel.grid.minor = element_blank(),
-              panel.border = element_blank(), panel.background = element_blank())
+        scale_y_continuous(labels = dollar_format(suffix = "", prefix = "")) +mytheme
 
       if(Flip==TRUE) {return(pp+coord_flip())} else {return(pp)}
     }
     else
-      if(!is.null(gp)) {
-        Tar_var = xx[,gp]
+      if(!is.null(target)) {
+        Tar_var = xx[,target]
         if(is.numeric(Tar_var)&length(unique(Tar_var))>5) {
           nlevel = length(unique(Xvar))
 
           if(is.null(col) & nlevel==2){fill_1 <- c("#5F9EA0", "#E1B378")} else
-            if(is.null(col) & nlevel>2){fill_1 <- sample(colors(),nlevel)}
+            if(is.null(col) & nlevel>2){fill_1 <- colors()[srswor(nlevel,length(colors()))==1]}
 
 
           tb <- table(Xvar)
@@ -138,16 +146,10 @@ ExpCatViz <- function(data,gp=NULL,fname=NULL,clim=10,col=NULL,margin=1,Page=NUL
           wrap_40 <- wrap_format(40)
           pp = ggplot(tbl, aes(y=Freq, x=Xvar,label = paste0(Freq,"%"))) +
             geom_bar( stat="identity", position="dodge",fill=fill_1)+xlab("")+
-            ylab("Percentage (%)")+ggtitle(wrap_40(paste(j)))+
-            geom_text(size = 3, position=position_dodge(width=0.9),vjust=0.8)+
+            ylab("Percentage (%)")+ggtitle(wrap_40(paste(chart_title," ",j)))+
+            geom_text(size = 4, position=position_dodge(width=0.9),vjust=0)+
             scale_x_discrete(labels = wrap_format(8))+
-            scale_y_continuous(labels = dollar_format(suffix = "%", prefix = ""))+
-            theme(axis.text.x=element_text(angle=90,vjust=.5,hjust=0.95,size=8,colour='grey20'),
-                  axis.text.y=element_text(vjust=.5,size=8,colour='grey20'),
-                  plot.title = element_text(hjust = 0.5, face = "bold", colour = "#5F9EA0", size = 12))+
-            theme(axis.line = element_line(size=1, colour = "black"),
-                  panel.grid.major = element_line(colour = "#d3d3d3",linetype = "dashed"), panel.grid.minor = element_blank(),
-                  panel.border = element_blank(), panel.background = element_blank())
+            scale_y_continuous(labels = dollar_format(suffix = "%", prefix = ""))+mytheme
 
           if(Flip==TRUE) {return(pp+coord_flip())} else {return(pp)}
 
@@ -161,17 +163,11 @@ ExpCatViz <- function(data,gp=NULL,fname=NULL,clim=10,col=NULL,margin=1,Page=NUL
           )
           pp = ggplot(tbl, aes(fill=Yvar, y=Freq, x=Xvar,label = paste0(Freq,"%"))) +
             geom_bar( stat="identity", position="dodge")+xlab("")+
-            ylab("Percentage (%)")+ggtitle(wrap_40(paste(j," vs ",gp)))+
-            geom_text(size = 3, position=position_dodge(width=0.9),vjust=0.8)+
+            ylab("Percentage (%)")+ggtitle(wrap_40(paste(chart_title," ",j," vs ",target,"[Target]")))+
+            geom_text(size = 4, position=position_dodge(width=0.9),vjust=0)+
             scale_x_discrete(labels = wrap_format(8))+
             scale_y_continuous(labels = dollar_format(suffix = "%", prefix = ""))+
-            scale_fill_manual("Target",values=fill_1)+
-            theme(axis.text.x=element_text(angle=90,vjust=.5,hjust=0.95,size=8,colour='grey20'),
-                  axis.text.y=element_text(vjust=.5,hjust=0.95,size=8,colour='grey20'),
-                  legend.position="top",plot.title = element_text(hjust = 0.5, face = "bold", colour = "#5F9EA0", size = 12))+
-            theme(axis.line = element_line(size=1, colour = "black"),
-                  panel.grid.major = element_line(colour = "#d3d3d3",linetype = "dashed"), panel.grid.minor = element_blank(),
-                  panel.border = element_blank(), panel.background = element_blank())
+            scale_fill_manual("Target",values=fill_1)+mytheme
 
           if(Flip==TRUE) {return(pp+coord_flip())} else {return(pp)}
 
@@ -180,9 +176,10 @@ ExpCatViz <- function(data,gp=NULL,fname=NULL,clim=10,col=NULL,margin=1,Page=NUL
     else
     {
       nlevel = length(unique(Xvar))
-
-      if(is.null(col) & nlevel==2){fill_1 <- c("#5F9EA0", "#E1B378")} else
-        if(is.null(col) & nlevel>2){fill_1 <- sample(colors(),nlevel)}
+      if(!is.null(col) & length(col)==1){fill_1 <- col} else
+        if(!is.null(col) & length(col)==2 & nlevel==2){fill_1 <- col} else
+          if(is.null(col) & nlevel==2){fill_1 <- c("#5F9EA0", "#E1B378")} else
+            if(is.null(col) & nlevel>2){fill_1 <- colors()[srswor(nlevel,length(colors()))==1]}
 
 
       tb <- table(Xvar)
@@ -191,16 +188,12 @@ ExpCatViz <- function(data,gp=NULL,fname=NULL,clim=10,col=NULL,margin=1,Page=NUL
 
       pp = ggplot(tbl, aes(y=Freq, x=Xvar,label = paste0(Freq,"%"))) +
         geom_bar( stat="identity", position="dodge",fill=fill_1)+xlab("")+
-        ylab("Percentage (%)")+ggtitle(wrap_40(paste(j)))+
-        geom_text(size = 3, position=position_dodge(width=0.9),vjust=0.8)+
+        ylab("Percentage (%)")+ggtitle(wrap_40(paste(chart_title," ",j)))+
+        geom_text(size = 4, position=position_dodge(width=0.9),vjust=0)+
         scale_x_discrete(labels = wrap_format(8))+
-        scale_y_continuous(labels = dollar_format(suffix = "%", prefix = ""))+
-        theme(axis.text.x=element_text(angle=90,vjust=.5,hjust=0.95,size=8,colour='grey20'),
-              axis.text.y=element_text(vjust=.5,hjust=0.95,size=8,colour='grey20'),
-              plot.title = element_text(hjust = 0.5, face = "bold", colour = "#5F9EA0", size = 12))+
-        theme(axis.line = element_line(size=1, colour = "black"),
-              panel.grid.major = element_line(colour = "#d3d3d3",linetype = "dashed"), panel.grid.minor = element_blank(),
-              panel.border = element_blank(), panel.background = element_blank())
+        scale_y_continuous(labels = dollar_format(suffix = "%", prefix = ""))+mytheme
+
+
 
       if(Flip==TRUE) {return(pp+coord_flip())} else {return(pp)}
 
@@ -240,6 +233,3 @@ ExpCatViz <- function(data,gp=NULL,fname=NULL,clim=10,col=NULL,margin=1,Page=NUL
   }
 
 }
-
-
-