+++ R/fn_exp_numeric.R
@@ -83,6 +83,7 @@
 #' ExpNumStat(mtcars,by="GA",gp="gear",Qnt=c(0.1,0.2),MesofShape=2,
 #' Outlier=FALSE,round=2,dcast = TRUE,val = "IQR")
 ##' @author dubrangala
+##' @importFrom stats quantile median IQR var reorder sd cor
 ##' @export ExpNumStat