+++ R/fn_exp_report.R
@@ -23,12 +23,8 @@
 ##' IF Target = continuous, then it will genreate EDA report including univariate and multivarite summary statistics with correlation.
 ##'
 ##' IF Target = categorical, then it will genreate EDA report including univariate and multivarite summary statistics with chi-square, Information values.
-##' @examples
-##' # Overview of the data
-##' ExpReport(mtcars,Template=NULL,Target=NULL,label=NULL,op_file="Myreport.html",
-##' op_dir=getwd(),sc=2,sn=2,Rc=NULL)
-##' @importFrom knitr kable
 ##' @importFrom rmarkdown render
+##' @importFrom utils browseURL
 ##' @export ExpReport
 
 ExpReport<-function (data,Template=NULL,Target=NULL,label=NULL,op_file=NULL,op_dir=getwd(),sc=NULL,sn=NULL,Rc=NULL)